+++ packages/plugins/eslint-plugin-react-x/src/utils/index.ts
@@ -1,6 +1,2 @@
 export * from "./constants";
 export * from "./create-rule";
-export * from "./is-from-hook-call";
-export * from "./is-from-use-state-call";
-export * from "./is-set-function-call";
-export * from "./is-variable-declarator-from-hook-call";
