--- packages/plugins/eslint-plugin-react-x/src/utils/is-variable-declarator-from-hook-call.ts
@@ -1,29 +1,0 @@
-import type { TSESTree } from "@typescript-eslint/types";
-import * as ER from "@eslint-react/core";
-import { AST_NODE_TYPES as T } from "@typescript-eslint/types";
-
-type VariableDeclaratorFromHookCall = {
-  id: TSESTree.Identifier;
-  init: { callee: TSESTree.Identifier | TSESTree.MemberExpression } & TSESTree.CallExpression;
-} & TSESTree.VariableDeclarator;
-
-export function isVariableDeclaratorFromHookCall(node: TSESTree.Node): node is VariableDeclaratorFromHookCall {
-  if (node.type !== T.VariableDeclarator) {
-    return false;
-  }
-  if (node.id.type !== T.Identifier) {
-    return false;
-  }
-  if (node.init?.type !== T.CallExpression) {
-    return false;
-  }
-  switch (node.init.callee.type) {
-    case T.Identifier:
-      return ER.isReactHookName(node.init.callee.name);
-    case T.MemberExpression:
-      return node.init.callee.property.type === T.Identifier
-        && ER.isReactHookName(node.init.callee.property.name);
-    default:
-      return false;
-  }
-}
